+++ app/views/admin/shared/v3/_pagy.html.haml
@@ -1,20 +1,20 @@
 %nav.pagy_nav.pagination{"aria-label" => "pager", :role => "navigation"}
   - if pagy.prev
-    %a.page.previous{ href: "#", "data-reflex": reflex, "data-page": pagy.prev || 1, "aria-label": "previous"}
+    %a.page.prev{ href: "#", "data-action": action, "data-page": pagy.prev || 1, "aria-label": "previous"}
       %i.icon-chevron-left
   - else
     %span.page.prev.disabled
       %i.icon-chevron-left
   - pagy.series.each do |item|       # series example: [1, :gap, 7, 8, "9", 10, 11, :gap, 36]
     - if item.is_a?(Integer)         # page link
-      %a.page{ href: "#", "data-reflex": reflex, "data-page": item, "aria-label": "page #{item}"}
+      %a.page{ href: "#", "data-action": action, "data-page": item, "aria-label": "page #{item}"}
         = item
     - elsif item.is_a?(String)       # current page
       %span.page.current= item
     - elsif item == :gap             # page gap
       %span.page.gap &hellip;
   - if pagy.next
-    %a.page.next{ href: "#", "data-reflex": reflex, "data-page": pagy.next || pagy.last, "aria-label": "next"}
+    %a.page.next{ href: "#", "data-action": action, "data-page": pagy.next || pagy.last, "aria-label": "next"}
       %i.icon-chevron-right
   - else
     %span.page.next.disabled
